pyproject.toml
[project]
name = "lanesentry"
version = "0.5.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cnocr>=2.3.1",
    "numpy>=2.2.6",
    "onnxruntime-gpu>=1.22.0",
    "opencv-python>=4.11.0.86",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
